Module Re2.Exceptions

exception Regex_no_such_subpattern of int * int

Regex_no_such_subpattern (n, max) means n was requested but only max subpatterns are defined (so max - 1 is the highest valid index)

exception Regex_no_such_named_subpattern of string * string

Regex_no_such_named_subpattern (name, pattern)

exception Regex_match_failed of string

Regex_match_failed pattern

exception Regex_submatch_did_not_capture of string * int

Regex_submatch_did_not_capture (s, i) means the ith subpattern in the regex compiled from s did not capture a substring.

exception Regex_compile_failed of string

the string is the C library's error message, generally in the form of "(human-readable error): (piece of pattern that did not compile)"

exception Regex_rewrite_template_invalid of string * string

Regex_rewrite_template_invalid (template, error_msg)