Module Re2.Stable

module V2 : sig ... end

V2 serializes, compares and hashes the pattern and the all currently known options, except max_mem. If Re2 gained a new option, we would have to mint a V3.t as the V2 format would not support this option. How we support such an upgrade (raise, drop the option, smash the tree) will have to considered then.

module V1_no_options : sig ... end

V1_no_options is the legacy serialization: pattern only, options are lost.